chore: move training related hooks to train model directory
