+++ application/ui/src/features/models/train-model/api/use-get-training-devices.ts
